requirements.txt
langchain==0.1.17
langchain-community>=0.0.36,<0.1
langchain-openai==0.1.8
faiss-cpu
openai>=1.0.0
pandas
python-dotenv
fastapi
uvicorn
tiktoken
pydantic
pydantic-settings